Added tests for OrderedLayer's consumers
