+++ src/trace/layers/column_leaf/tests.rs
@@ -178,146 +178,18 @@ fn no_double_drops_during_partial_abandonment() {
 
 #[cfg_attr(miri, ignore)]
 mod proptests {
-    use crate::trace::{
-        layers::{
-            column_leaf::{ColumnLeafConsumer, OrderedColumnLeaf, OrderedColumnLeafBuilder},
-            Builder, TupleBuilder,
+    use crate::{
+        trace::{
+            layers::{
+                column_leaf::{ColumnLeafConsumer, OrderedColumnLeafBuilder},
+                Builder, TupleBuilder,
+            },
+            Consumer,
         },
-        Consumer,
+        utils::tests::{orderings, ArtificialPanic, LimitedDrops, RandomlyOrdered},
     };
-    use proptest::{collection::vec, prelude::*};
-    use std::{
-        cell::{Cell, RefCell},
-        cmp::Ordering,
-        panic::{self, AssertUnwindSafe},
-        rc::Rc,
-        thread,
-    };
-
-    prop_compose! {
-        /// Generate a random vec of orderings
-        fn orderings(max_length: usize)
-            (orderings in vec(any::<Ordering>(), 0..=max_length))
-        -> Vec<Ordering> {
-            orderings
-        }
-    }
-
-    prop_compose! {
-        /// Generate a OrderedColumnLeaf
-        fn column_leaf(max_length: usize)
-            (length in 0..=max_length)
-            (length in length..=length, diffs in vec(any::<isize>(), length))
-        -> OrderedColumnLeaf<usize, isize> {
-            let mut builder = OrderedColumnLeafBuilder::with_capacity(length);
-            for (idx, diff) in diffs.into_iter().enumerate() {
-                builder.push_tuple((idx, diff));
-            }
-
-            builder.done()
-        }
-    }
-
-    #[derive(Debug)]
-    struct ArtificialPanic;
-
-    #[derive(Debug, Clone)]
-    struct RandomlyOrdered {
-        orderings: Rc<RefCell<Vec<Ordering>>>,
-    }
-
-    impl RandomlyOrdered {
-        pub fn new(orderings: Vec<Ordering>) -> Self {
-            Self {
-                orderings: Rc::new(RefCell::new(orderings)),
-            }
-        }
-    }
-
-    impl PartialEq for RandomlyOrdered {
-        fn eq(&self, _other: &Self) -> bool {
-            false
-        }
-    }
-
-    impl Eq for RandomlyOrdered {}
-
-    impl PartialOrd for RandomlyOrdered {
-        fn partial_cmp(&self, _other: &Self) -> Option<Ordering> {
-            match self.orderings.borrow_mut().pop() {
-                order @ Some(_) => order,
-                None => panic::panic_any(ArtificialPanic),
-            }
-        }
-    }
-
-    impl Ord for RandomlyOrdered {
-        fn cmp(&self, _other: &Self) -> Ordering {
-            match self.orderings.borrow_mut().pop() {
-                Some(order) => order,
-                None => panic::panic_any(ArtificialPanic),
-            }
-        }
-    }
-
-    #[derive(Clone)]
-    struct LimitedDrops<T> {
-        inner: T,
-        allowed_drops: Option<Rc<Cell<usize>>>,
-    }
-
-    impl<T> LimitedDrops<T> {
-        fn new(inner: T, allowed_drops: Option<Rc<Cell<usize>>>) -> Self {
-            Self {
-                inner,
-                allowed_drops,
-            }
-        }
-    }
-
-    impl<T> PartialEq for LimitedDrops<T>
-    where
-        T: PartialEq,
-    {
-        fn eq(&self, other: &Self) -> bool {
-            self.inner.eq(&other.inner)
-        }
-    }
-
-    impl<T: Eq> Eq for LimitedDrops<T> {}
-
-    impl<T> PartialOrd for LimitedDrops<T>
-    where
-        T: PartialOrd,
-    {
-        fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-            self.inner.partial_cmp(&other.inner)
-        }
-    }
-
-    impl<T> Ord for LimitedDrops<T>
-    where
-        T: Ord,
-    {
-        fn cmp(&self, other: &Self) -> Ordering {
-            self.inner.cmp(&other.inner)
-        }
-    }
-
-    impl<T> Drop for LimitedDrops<T> {
-        fn drop(&mut self) {
-            if let Some(allowed_drops) =
-                self.allowed_drops.as_ref().filter(|_| !thread::panicking())
-            {
-                let remaining_drops = allowed_drops.get();
-                if remaining_drops == 0 {
-                    panic::panic_any(ArtificialPanic);
-                } else {
-                    allowed_drops.set(remaining_drops - 1);
-                }
-            }
-        }
-    }
+    use proptest::prelude::*;
+    use std::{cell::Cell, rc::Rc};
 
     proptest! {
         #[test]
@@ -337,14 +209,7 @@ mod proptests {
             // Seek repeatedly to incur a panic within the comparison function
             let needle = Box::new(orderings);
             for _ in 0..100 {
-                let result = panic::catch_unwind(AssertUnwindSafe(|| consumer.seek_key(&needle)));
-
-                // Catch any of our artificially induced panics
-                if let Err(payload) = result {
-                    if !payload.is::<ArtificialPanic>() {
-                        panic::resume_unwind(payload);
-                    }
-                }
+                ArtificialPanic::catch(|| consumer.seek_key(&needle));
             }
         }
 
@@ -365,17 +230,10 @@ mod proptests {
 
             // We seek for a random key which should drop some elements and then we drop the
             // consumer which should drop all remaining elements
-            let result = panic::catch_unwind(AssertUnwindSafe(|| {
+            ArtificialPanic::catch(||  {
                 consumer.seek_key(&LimitedDrops::new(Box::new(needle), None));
                 drop(consumer);
-            }));
-
-            // Catch any of our artificially induced panics
-            if let Err(payload) = result {
-                if !payload.is::<ArtificialPanic>() {
-                    panic::resume_unwind(payload);
-                }
-            }
+            });
         }
     }
 }
